+++ packages/remix-dev/cli/codemod/run/transforms/types.ts
@@ -1,0 +1,8 @@
+import type { Answers } from "../../questions";
+
+export type TransformArgs = {
+  answers: Answers;
+  files: string[];
+  flags: { dry?: boolean; print?: boolean; runInBand?: boolean };
+};
+export type Transform = (args: TransformArgs) => void | Promise<void>;
